Fix the issue of error message not shown by the GUI interface
