+++ src/main/java/gopher/parser/Parser.java
@@ -7,8 +7,10 @@ import java.time.format.DateTimeParseException;
 import gopher.exception.EmptyTaskDescriptionException;
 import gopher.exception.MissingTokenException;
 import gopher.exception.UnknownCommandException;
+import gopher.task.Deadline;
+import gopher.task.Event;
 import gopher.task.Task;
-import gopher.ui.UI;
+import gopher.task.ToDo;
 
 /**
  * Groups the logic and functions for parsing input, command and data from
@@ -73,18 +75,107 @@ public class Parser {
      * @param command string input from user or file
      * @return task of the correct type with relevant input details
      */
-    public static Task parseCreateTaskCommand(String command) {
-        try {
-            return Task.of(command);
-        } catch (UnknownCommandException e) {
-            UI.printUnknownCommandWarning(e);
-        } catch (DateTimeParseException e) {
-            UI.printInvalidDateWarning();
-        } catch (EmptyTaskDescriptionException
-                 | MissingTokenException e) {
-            System.out.println(e.getMessage());
+    public static Task parseCreateTaskCommand(String command)
+            throws UnknownCommandException, DateTimeParseException,
+            EmptyTaskDescriptionException, MissingTokenException {
+        String[] tokens = command.split(" ");
+
+        String taskType = tokens[0];
+        if (!taskType.equalsIgnoreCase("todo")
+                && !taskType.equalsIgnoreCase("deadline")
+                && !taskType.equalsIgnoreCase("event")) {
+            throw new UnknownCommandException(taskType);
         }
-        return null;
+
+        if (tokens.length < 2) {
+            throw new EmptyTaskDescriptionException(taskType);
+        }
+
+        StringBuilder taskName = new StringBuilder();
+        if (taskType.equalsIgnoreCase("todo")) {
+            for (int i = 1; i < tokens.length; i++) {
+                taskName.append(tokens[i]);
+                if (i < tokens.length - 1) {
+                    taskName.append(" ");
+                }
+            }
+            return new ToDo(taskName.toString());
+        } else if (taskType.equalsIgnoreCase("deadline")) {
+            int byTokenIndex = -1;
+            for (int i = 1; i < tokens.length; i++) {
+                if (tokens[i].equalsIgnoreCase("/by")) {
+                    byTokenIndex = i;
+                }
+            }
+
+            if (byTokenIndex == -1) {
+                throw new MissingTokenException(taskType, "/by");
+            }
+
+            for (int i = 1; i < byTokenIndex; i++) {
+                taskName.append(tokens[i]);
+                if (i < byTokenIndex - 1) {
+                    taskName.append(" ");
+                }
+            }
+
+            StringBuilder dueDate = new StringBuilder();
+            for (int i = byTokenIndex + 1; i < tokens.length; i++) {
+                dueDate.append(tokens[i]);
+                if (i < tokens.length - 1) {
+                    dueDate.append(" ");
+                }
+            }
+
+            return new Deadline(taskName.toString(), dueDate.toString());
+        } else if (taskType.equalsIgnoreCase("event")) {
+            int fromTokenIndex = -1;
+            int toTokenIndex = -1;
+            for (int i = 1; i < tokens.length; i++) {
+                if (tokens[i].equalsIgnoreCase("/from")) {
+                    fromTokenIndex = i;
+                }
+                if (tokens[i].equalsIgnoreCase("/to")) {
+                    toTokenIndex = i;
+                }
+            }
+
+            if (fromTokenIndex == -1) {
+                throw new MissingTokenException(taskType, "/from");
+            }
+
+            if (toTokenIndex == -1) {
+                throw new MissingTokenException(taskType, "/to");
+            }
+
+            for (int i = 1; i < fromTokenIndex; i++) {
+                taskName.append(tokens[i]);
+                if (i < fromTokenIndex - 1) {
+                    taskName.append(" ");
+                }
+            }
+
+            StringBuilder startDate = new StringBuilder();
+            for (int i = fromTokenIndex + 1; i < toTokenIndex; i++) {
+                startDate.append(tokens[i]);
+                if (i < toTokenIndex - 1) {
+                    startDate.append(" ");
+                }
+            }
+
+            StringBuilder endDate = new StringBuilder();
+            for (int i = toTokenIndex + 1; i < tokens.length; i++) {
+                endDate.append(tokens[i]);
+                if (i < tokens.length - 1) {
+                    endDate.append(" ");
+                }
+            }
+            return new Event(taskName.toString(),
+                    startDate.toString(),
+                    endDate.toString());
+        } else {
+            throw new UnknownCommandException(taskType);
+        }
     }
 
     /**
