+++ src/main/java/gopher/ui/UI.java
@@ -9,133 +9,117 @@ import gopher.task.TaskList;
  */
 public class UI {
     /**
-     * Prints the Gopher Logo.
+     * Get the greeting message to show to the user
      */
-    public static void printGopherLogo() {
-        System.out.println("""
-                  ____             _
-                 / ___| ___  _ __ | |__   ___ _ __
-                | |  _ / _ \\| '_ \\| '_ \\ / _ \\ '__|
-                | |_| | (_) | |_) | | | |  __/ |
-                 \\____|\\___/| .__/|_| |_|\\___|_|
-                            |_|
-                """);
+    public static String getGreetMessage() {
+        return "Hi, welcome to Gopher!\n"
+                + "Here are the available commands:\n"
+                + """
+                1. todo/deadline/event - Create event of these types
+                2. mark/unmark X - Mark X-th task as done/not done
+                3. find [keyword] - Find matching tasks based on keywords
+                4. delete X - Delete X-th task
+                5. list - List out the current tasks
+                6. bye - End the interaction
+                """;
     }
 
     /**
-     * Prints the horizontal separator.
-     */
-    public static void printHorizontalSeparator() {
-        System.out.println("==================================================");
-    }
-
-    /**
-     * Prints the greeting message.
-     */
-    public static void printGreetMessage() {
-        printGopherLogo();
-        System.out.println("Hello! I am Gopher.\nWhat can I do for you?");
-    }
-
-    /**
-     * Prints the given TaskList for task visualization.
+     * Get task visualization for the given TaskList.
      *
      * @param tasks TaskList object to be printed
      */
-    public static void printTaskList(TaskList tasks) {
-        System.out.println(String.format("You currently have %d %s in the list\n",
+    public static String getTaskListMessage(TaskList tasks) {
+        return String.format("You currently have %d %s in the list\n%s",
                 tasks.getSize(),
-                tasks.getSize() == 1 ? "task" : "tasks"));
-        System.out.println(tasks);
+                tasks.getSize() == 1 ? "task" : "tasks",
+                tasks);
     }
 
     /**
-     * Print the matched tasks from the given TaskList.
+     * Get the matched tasks message based on the given matched TaskList.
      *
      * @param tasks TaskList object containing all the matched tasks
      */
-    public static void printMatchedTasks(TaskList tasks) {
-        System.out.println(String.format("Found %d matching %s in your list:",
+    public static String getMatchedTasksMessage(TaskList tasks) {
+        return String.format("Found %d matching %s in your list:\n%s",
                 tasks.getSize(),
-                tasks.getSize() == 1 ? "task" : "tasks"
-        ));
-        System.out.println(tasks);
+                tasks.getSize() == 1 ? "task" : "tasks",
+                tasks
+        );
     }
 
     /**
-     * Prints the message when TaskList successfully add in a new task.
+     * Get the message when TaskList successfully add in a new task.
      *
-     * @param taskList TaskList object that accepts the new task
      * @param task     task to be added
      */
-    public static void printAddTaskMessage(TaskList taskList, Task task) {
-        System.out.println("Got it! I have added this task:\n" + task);
+    public static String getAddTaskMessage(Task task) {
+        return "Got it! I have added this task:\n" + task;
     }
 
     /**
-     * Prints the message when a given task is successfully mark as done.
+     * Get the message when a given task is successfully mark as done.
      *
      * @param task task to be marked as done
      */
-    public static void printMarkAsDoneMessage(Task task) {
-        System.out.println("Nice! I've marked this task as done:");
-        System.out.println(task);
+    public static String getMarkAsDoneMessage(Task task) {
+        return "Nice! I've marked this task as done:\n" + task;
     }
 
     /**
-     * Prints the message when a given task is successfully mark as not done.
+     * Get the message when a given task is successfully mark as not done.
      *
      * @param task task to be marked as not done
      */
-    public static void printMarkAsUndoneMessage(Task task) {
-        System.out.println("Ok, I've marked this task as not done yet:");
-        System.out.println(task);
+    public static String getMarkAsUndoneMessage(Task task) {
+        return "Ok, I've marked this task as not done yet:\n" + task;
     }
 
     /**
-     * Prints the message when TaskList successfully delete a task.
+     * Get the message when TaskList successfully delete a task.
      *
-     * @param taskList TaskList in which the task needs to be deleted
      * @param task     task to be deleted
      */
-    public static void printDeleteTaskMessage(TaskList taskList, Task task) {
-        System.out.println("Noted. I've removed this task:");
-        System.out.println(task);
+    public static String getDeleteTaskMessage(Task task) {
+        return "Noted. I've removed this task:\n" + task;
     }
 
     /**
-     * Prints the exit message.
+     * Get the exit message.
      */
-    public static void printExitMessage() {
-        System.out.println("Bye. Hope to see you again soon!");
+    public static String getExitMessage() {
+        return "Bye. Hope to see you again soon!";
     }
 
     /**
-     * Prints the helping message when user encounters an UnknownCommandException.
+     * Get the helping message when user encounters an UnknownCommandException.
      *
      * @param e UnknownCommandException thrown by the system
      */
-    public static void printUnknownCommandWarning(UnknownCommandException e) {
-        System.out.println(e.getMessage());
-        System.out.println("""
-                Current Supported Commands:
-                    1. todo - Create a ToDo Task
-                    2. deadline - Create a Deadline Task
-                    3. event - Create an Event Task
-                    4. mark - Mark a task as done
-                    5. unmark - Mark a task as not done
-                    6. bye - Exit the chatbot
+    public static String getUnknownCommandWarning(UnknownCommandException e) {
+        return e.getMessage()
+                +
+                """
+                    Current Supported Commands:
+                        1. todo - Create a ToDo Task
+                        2. deadline - Create a Deadline Task
+                        3. event - Create an Event Task
+                        4. mark - Mark a task as done
+                        5. unmark - Mark a task as not done
+                        6. find - Find tasks based on keywords
+                        7. bye - Exit the chatbot
 
-                Note: The command is case-insensitive, as long as the
-                input characters match, the chatbot would be able to
-                respond to the given command""");
+                    Note: The command is case-insensitive,
+                    as long as the input characters match,
+                    the chatbot would be able to respond to the given command""";
     }
 
     /**
-     * Prints the helping message when user inputs invalid date.
+     * Get the helping message when user inputs invalid date.
      */
-    public static void printInvalidDateWarning() {
-        System.out.println("Oops...Seems like you have passed in invalid date...\n"
+    public static String getInvalidDateWarning() {
+        return ("Oops...Seems like you have passed in invalid date...\n"
                 + "Please check if your date is of the correct value and format\n"
                 + "Current accepted date format: YYYY-MM-DD or YYYY-MM-DD hh:mm");
     }
